fixed useMemo diffing bug. falsy must indicate loading, not status values.
